+++ CHANGELOG.md
@@ -21,6 +21,7 @@ The types of changes are:
 - Added a `reinitialize` method to Fides.js [#4812](https://github.com/ethyca/fides/pull/4812)
 - Added carets to collapsible sections in the overlay modal [#4793](https://github.com/ethyca/fides/pull/4793)
 - Added erasure support for OpenWeb [#4735](https://github.com/ethyca/fides/pull/4735)
+- Added support for configuration of pre-approval webhooks [#4795](https://github.com/ethyca/fides/pull/4795)
 
 ### Changed
 - Removed the Celery startup banner from the Fides worker logs [#4814](https://github.com/ethyca/fides/pull/4814)